extract computeDeploysCheckpoint function
